[cd-arranges/_lossless] Fix mono/stereo mismatch...
...in th095_03

For both the lossy and lossless patches, the "intro" file is mono, whereas the "loop" file is stereo. This prompts a warning from thcrap when th9.5 is launched.

This commit fixes this issue by replacing both mono files with stereo versions of the same audio (made by duplicating the mono track).